deps: Downgrade spl crates for upcoming yank
Problem

New projects using spl crates are having a hard time because of the flexible
dependency on v2 or v1 of the Solana crates -- cargo picks up the highest
version of the dependency if there's no lockfile, and downstream users are
left with confusing error messages about multiple versions of Solana crates
present in the project.

To make it easier for new projects, we are yanking a lot of SPL crates that
have the flexible dependency, and spl-type-length-value 0.3.1 is one of those.

This crate is the only dependency that explicitly uses it.

Summary of Changes

Downgrade spl-token-2022 to v1 and spl-type-length-value to v0.3.0 to avoid
using it.

More information at https://github.com/solana-labs/solana-program-library/issues/6897